requirements.txt
yfinance
ddgs
langgraph
langchain-openai
python-dotenv